refactor: phase 3 sync domain modules and snapshot contracts
