+++ api/chat/tools/types.ts
@@ -5,8 +5,6 @@
  * supporting both server-side and client-side tool execution.
  */
 
-import type { Contact } from "../../../src/utils/contacts.js";
-
 // Central list of supported theme IDs for tool validation
 export const THEME_IDS = ["system7", "macosx", "xp", "win98"] as const;
 export type ThemeId = typeof THEME_IDS[number];
@@ -244,43 +242,15 @@ export interface SettingsInput {
   checkForUpdates?: boolean;
 }
 
-// Sticky colors
-export const STICKY_COLORS = ["yellow", "blue", "green", "pink", "purple", "orange"] as const;
-export type StickyColor = typeof STICKY_COLORS[number];
-
-// Stickies control actions
-export const STICKIES_ACTIONS = ["list", "create", "update", "delete", "clear"] as const;
-export type StickiesAction = typeof STICKIES_ACTIONS[number];
-
-// Stickies control input
-export interface StickiesControlInput {
-  action: StickiesAction;
-  id?: string;
-  content?: string;
-  color?: StickyColor;
-  position?: { x: number; y: number };
-  size?: { width: number; height: number };
-}
-
-// Stickies control output
-export interface StickiesControlOutput {
-  success: boolean;
-  message: string;
-  notes?: Array<{
-    id: string;
-    content: string;
-    color: StickyColor;
-    position: { x: number; y: number };
-    size: { width: number; height: number };
-  }>;
-  note?: {
-    id: string;
-    content: string;
-    color: StickyColor;
-    position: { x: number; y: number };
-    size: { width: number; height: number };
-  };
-}
+export {
+  STICKY_COLORS,
+  STICKIES_ACTIONS,
+  type StickyColor,
+  type StickiesAction,
+  type StickiesControlInput,
+  type StickiesControlOutput,
+  type StickyNoteToolRecord,
+} from "@ryos/shared/actions/stickies";
 
 // Infinite Mac control actions
 export const INFINITE_MAC_ACTIONS = [
@@ -426,111 +396,28 @@ export interface TvControlOutput {
 // Calendar Control Types
 // ============================================================================
 
-export const CALENDAR_ACTIONS = ["list", "create", "update", "delete", "listTodos", "createTodo", "toggleTodo", "deleteTodo"] as const;
-export type CalendarAction = typeof CALENDAR_ACTIONS[number];
-
-export const CALENDAR_COLORS = ["blue", "red", "green", "orange", "purple"] as const;
-export type CalendarColor = typeof CALENDAR_COLORS[number];
-
-export interface CalendarControlInput {
-  action: CalendarAction;
-  id?: string;
-  title?: string;
-  date?: string; // YYYY-MM-DD
-  startTime?: string; // HH:MM
-  endTime?: string; // HH:MM
-  color?: CalendarColor;
-  notes?: string;
-  completed?: boolean;
-  calendarId?: string;
-}
-
-export interface CalendarControlOutput {
-  success: boolean;
-  message: string;
-  events?: Array<{
-    id: string;
-    title: string;
-    date: string;
-    startTime?: string;
-    endTime?: string;
-    color: string;
-    notes?: string;
-  }>;
-  event?: {
-    id: string;
-    title: string;
-    date: string;
-    startTime?: string;
-    endTime?: string;
-    color: string;
-    notes?: string;
-  };
-  todos?: Array<{
-    id: string;
-    title: string;
-    completed: boolean;
-    dueDate: string | null;
-    calendarId: string;
-  }>;
-  todo?: {
-    id: string;
-    title: string;
-    completed: boolean;
-    dueDate: string | null;
-    calendarId: string;
-  };
-}
+export {
+  CALENDAR_ACTIONS,
+  CALENDAR_COLORS,
+  type CalendarAction,
+  type CalendarColor,
+  type CalendarControlInput,
+  type CalendarControlOutput,
+  type CalendarEventToolRecord,
+  type CalendarTodoToolRecord,
+} from "@ryos/shared/actions/calendar";
 
 // ============================================================================
 // Contacts Control Types
 // ============================================================================
 
-export const CONTACT_ACTIONS = ["list", "get", "create", "update", "delete"] as const;
-export type ContactsAction = typeof CONTACT_ACTIONS[number];
-
-export interface ContactsControlInput {
-  action: ContactsAction;
-  id?: string;
-  query?: string;
-  displayName?: string;
-  firstName?: string;
-  lastName?: string;
-  nickname?: string;
-  organization?: string;
-  title?: string;
-  notes?: string;
-  emails?: string[];
-  phones?: string[];
-  urls?: string[];
-  addresses?: string[];
-  birthday?: string | null;
-  telegramUsername?: string | null;
-  telegramUserId?: string | null;
-}
-
-export interface ContactToolRecord {
-  id: string;
-  displayName: string;
-  organization: string;
-  title: string;
-  emails: string[];
-  phones: string[];
-  urls: string[];
-  addresses: string[];
-  telegramUsername: string | null;
-  telegramUserId: string | null;
-  birthday: string | null;
-  notes?: string | null;
-  summary?: string | null;
-}
-
-export interface ContactsControlOutput {
-  success: boolean;
-  message: string;
-  contacts?: ContactToolRecord[];
-  contact?: ContactToolRecord | null;
-}
+export {
+  CONTACT_ACTIONS,
+  type ContactsAction,
+  type ContactsControlInput,
+  type ContactsControlOutput,
+  type ContactToolRecord,
+} from "@ryos/shared/actions/contacts";
 
 // ============================================================================
 // Documents Control Types
@@ -736,53 +623,8 @@ export interface WebFetchOutput {
 // App State Types (for server-side calendar/stickies executors)
 // ============================================================================
 
-export interface CalendarSnapshotData {
-  events: Array<{
-    id: string;
-    title: string;
-    date: string;
-    startTime?: string;
-    endTime?: string;
-    color: string;
-    calendarId?: string;
-    notes?: string;
-    createdAt: number;
-    updatedAt: number;
-  }>;
-  calendars: Array<{
-    id: string;
-    name: string;
-    color: string;
-    visible: boolean;
-  }>;
-  todos: Array<{
-    id: string;
-    title: string;
-    completed: boolean;
-    dueDate: string | null;
-    calendarId: string;
-    createdAt: number;
-  }>;
-  deletedEventIds?: Record<string, string>;
-  deletedCalendarIds?: Record<string, string>;
-  deletedTodoIds?: Record<string, string>;
-}
-
-export interface StickiesSnapshotData {
-  notes: Array<{
-    id: string;
-    content: string;
-    color: string;
-    position: { x: number; y: number };
-    size: { width: number; height: number };
-    createdAt: number;
-    updatedAt: number;
-  }>;
-  deletedNoteIds?: Record<string, string>;
-}
-
-export interface ContactsSnapshotData {
-  contacts: Contact[];
-  myContactId?: string | null;
-  deletedContactIds?: Record<string, string>;
-}
+export type {
+  CalendarSnapshotData,
+  StickiesSnapshotData,
+  ContactsSnapshotData,
+} from "@ryos/shared/contracts/sync-snapshots";
