+++ shared/index.ts
@@ -2,3 +2,7 @@ export * from "./contracts/ai-models";
 export * from "./contracts/chat-rooms";
 export * from "./contracts/listen";
 export * from "./contracts/songs";
+export * from "./contracts/sync-snapshots";
+export * from "./actions/calendar";
+export * from "./actions/contacts";
+export * from "./actions/stickies";
